refactor : InputView의 자동차 이름 중복 체크 로직을 RacingGame의 생성자로 이동
