+++ src/main/java/racingcar/model/RacingGame.java
@@ -1,20 +1,40 @@
 package racingcar.model;
 
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 public class RacingGame {
+    public static final int MAX_TRY_COUNT_LIMIT = 99;
 
     private final NumberGenerator numberGenerator;
-    private final int maxTryCount;
     private final Cars cars;
+    private final int maxTryCount;
     private int tryCount;
 
     public RacingGame(int maxTryCount, NumberGenerator numberGenerator, List<String> names) {
+        if (hasDuplicatedName(names)) {
+            throw new IllegalArgumentException("자동차 이름은 유일한 이름이어야 합니다.");
+        }
+
+        if (isInvalidTryCount(maxTryCount)) {
+            throw new IllegalArgumentException("시도 회수는 " + MAX_TRY_COUNT_LIMIT + "이하의 숫자여야합니다.");
+        }
+
         this.numberGenerator = numberGenerator;
         this.cars = createCars(names);
         this.maxTryCount = maxTryCount;
         this.tryCount = 0;
+    }
+
+    private static boolean isInvalidTryCount(int maxTryCount) {
+        return maxTryCount > MAX_TRY_COUNT_LIMIT;
+    }
+
+    private boolean hasDuplicatedName(List<String> names) {
+        Set<String> set = new HashSet<>(names);
+        return set.size() != names.size();
     }
 
     private Cars createCars(List<String> names) {
