+++ src/main/java/racingcar/view/InputView.java
@@ -2,15 +2,13 @@ package racingcar.view;
 
 import java.io.InputStream;
 import java.io.PrintStream;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Scanner;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 public class InputView {
 
-    public static final int MAX_TRY_COUNT = 99;
+    public static final String NAME_SPLIT_REGEX = ",";
     private final PrintStream printStream;
     private final Scanner scanner;
 
@@ -21,27 +19,12 @@ public class InputView {
 
     public List<String> inputCarNames() {
         printStream.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
-        List<String> names = getCarNames();
-
-        if (hasDuplicatedName(names)) {
-            throw new IllegalArgumentException("자동차 이름은 유일한 이름이어야 합니다.");
-        }
-
-        return names;
-    }
-
-    private List<String> getCarNames() {
         String names = scanner.nextLine();
         return trim(split(names));
     }
 
-    private boolean hasDuplicatedName(List<String> names) {
-        Set<String> set = new HashSet<>(names);
-        return set.size() != names.size();
-    }
-
     private List<String> split(String names) {
-        return List.of(names.split(","));
+        return List.of(names.split(NAME_SPLIT_REGEX));
     }
 
     private List<String> trim(List<String> names) {
@@ -52,16 +35,14 @@ public class InputView {
 
     public int inputMaxTryCount() {
         printStream.println("시도할 회수는 몇회인가요?");
-        String value = scanner.nextLine();
-
-        if (isInvalidTryCount(value)) {
-            throw new IllegalArgumentException("시도 회수는 " + MAX_TRY_COUNT + "이하의 숫자여야합니다.");
+        String maxTryCount = scanner.nextLine();
+        if(!isInteger(maxTryCount)){
+            throw new IllegalArgumentException("시도할 횟수는 정수여야합니다.");
         }
-
-        return Integer.parseInt(value);
+        return Integer.parseInt(maxTryCount);
     }
 
-    private static boolean isInvalidTryCount(String value) {
-        return !(value.matches("\\d+") && Integer.parseInt(value) <= MAX_TRY_COUNT);
+    public boolean isInteger(String value){
+        return value.matches("[-+]?\\d+");
     }
 }
